requirements_dev.txt
-r requirements.txt
isort
black
pytest
pytest-playwright